third_party/WebKit/LayoutTests/bluetooth/idl/idl-Bluetooth.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

test(() => {
  assert_throws(null, () => new Bluetooth(),
                'the constructor should not be callable with "new"');
  assert_throws(null, () => Bluetooth(),
                'the constructor should not be callable');

  // Bluetooth implements BluetoothDiscovery;
  assert_true('requestDevice' in navigator.bluetooth);
}, 'Bluetooth IDL test');
</script>
